Fixed some bugs with the 'default-html-compiler'.

+++ scripts-ts/default/default-inline-script-compiler.ts
@@ -48,16 +48,11 @@ async function addHtml(dev: Dev, dirName: string, html: string): Promise<void> {
   await fs.writeFile(path.join(dev.args.out, dirName, dirName + '.html'), html);
 }
 
-async function findFile(
-  dev: Dev,
-  dirName: string,
-  ext: string
-): Promise<{ filePath: string; fileContent: string }> {
+async function findFile(dev: Dev, dirName: string, ext: string): Promise<{ filePath: string; fileContent: string }> {
   const inDirName = path.join(dev.args.in, dirName);
 
   let filePath =
-    (await dev.getFile(dev.args, path.join(inDirName, dirName + ext))) ??
-    path.join(inDirName, dirName + ext);
+    (await dev.getFile(dev.args, path.join(inDirName, dirName + ext))) ?? path.join(inDirName, dirName + ext);
 
   if (!fs.existsSync(filePath))
     return {
@@ -68,12 +63,7 @@ async function findFile(
   let fileContent = await fs.readFile(filePath);
   fileContent = fileContent.toString();
 
-  const res = await dev.pluginPipe(
-    'pipeFile',
-    { filePath, fileContent },
-    dev.args,
-    false
-  );
+  const res = await dev.pluginPipe('pipeFile', { filePath, fileContent }, dev.args, false);
 
   filePath = res.filePath ?? filePath;
   fileContent = res.fileContent ?? fileContent;
@@ -81,61 +71,33 @@ async function findFile(
   return { filePath, fileContent };
 }
 
-function getStyle(
-  dev: Dev,
-  dirName: string
-): Promise<{ filePath: string; fileContent: string }> {
+function getStyle(dev: Dev, dirName: string): Promise<{ filePath: string; fileContent: string }> {
   return findFile(dev, dirName, '.css');
 }
 
-function getScript(
-  dev: Dev,
-  dirName: string
-): Promise<{ filePath: string; fileContent: string }> {
+function getScript(dev: Dev, dirName: string): Promise<{ filePath: string; fileContent: string }> {
   return findFile(dev, dirName, '.js');
 }
 
-function getHTML(
-  dev: Dev,
-  dirName: string
-): Promise<{ filePath: string; fileContent: string }> {
+function getHTML(dev: Dev, dirName: string): Promise<{ filePath: string; fileContent: string }> {
   return findFile(dev, dirName, '.inline-script');
 }
 
-async function inlineScriptCompiler(
-  dev: Dev,
-  event: string,
-  filePath: string,
-  dirName: string
-): Promise<boolean> {
-  const style = (await getStyle(dev, dirName)).fileContent;
-  const script = (await getScript(dev, dirName)).fileContent;
+async function inlineScriptCompiler(dev: Dev, event: string, filePath: string, dirName: string): Promise<boolean> {
+  const style = (await getStyle(dev, dirName)).fileContent ?? '';
+  const script = (await getScript(dev, dirName)).fileContent ?? '';
 
   if (event === 'unlink') {
-    await fs.writeFile(
-      path.join(dev.args.out, dirName, dirName + '.css'),
-      style
-    );
-    await fs.writeFile(
-      path.join(dev.args.out, dirName, dirName + '.js'),
-      script
-    );
+    await fs.writeFile(path.join(dev.args.out, dirName, dirName + '.css'), style);
+    await fs.writeFile(path.join(dev.args.out, dirName, dirName + '.js'), script);
     await fs.unlink(path.join(dev.args.out, dirName, dirName + '.html'));
     return true;
   }
 
   const html = (await getHTML(dev, dirName)).fileContent;
 
-  if (style !== null)
-    await fs.unlink(
-      path.join(dev.args.out, dirName, dirName + '.css'),
-      (err) => {}
-    );
-  if (script !== null)
-    await fs.unlink(
-      path.join(dev.args.out, dirName, dirName + '.js'),
-      (err) => {}
-    );
+  if (style !== null) await fs.unlink(path.join(dev.args.out, dirName, dirName + '.css'), (err) => {});
+  if (script !== null) await fs.unlink(path.join(dev.args.out, dirName, dirName + '.js'), (err) => {});
 
   await addTemplate(dev, dirName);
   await addHtml(dev, dirName, generateHTML(html, style, script));
@@ -143,10 +105,7 @@ async function inlineScriptCompiler(
   return true;
 }
 
-async function updateInlineScriptFile(
-  dev: Dev,
-  inlineScriptPath: string
-): Promise<boolean> {
+async function updateInlineScriptFile(dev: Dev, inlineScriptPath: string): Promise<boolean> {
   dev.eventHandler('change', inlineScriptPath);
 
   return true;
@@ -157,11 +116,7 @@ let plugin: DunyaPlugin = {
 };
 
 plugin.name = 'default-inline-script-compiler';
-plugin.beforeWatchEventHalter = async function (
-  dev: Dev,
-  event: string,
-  filePath: string
-): Promise<boolean> {
+plugin.beforeWatchEventHalter = async function (dev: Dev, event: string, filePath: string): Promise<boolean> {
   const ext = path.extname(filePath);
 
   const dirName = path.dirname(filePath);
@@ -173,8 +128,7 @@ plugin.beforeWatchEventHalter = async function (
 
   const inlineScriptFile = await findFile(dev, dirName, '.inline-script');
   const inlineScriptPath = inlineScriptFile.filePath;
-  if (inlineScriptPath !== null)
-    return await updateInlineScriptFile(dev, inlineScriptPath);
+  if (inlineScriptPath !== null) return await updateInlineScriptFile(dev, inlineScriptPath);
 
   return false;
 };
